Fix postgres issues and start from scratch
I ran into major issues while testing the basic rails app after
creation. These had to do with the postgres user as well as the
locale on the VM. I researched and fixed these issues with the
resources in fixing-pg-issues.txt. I then destroyed the VM and
started over with the new project.